chore: working on player loading and a few other things.
Made prompt a little nicer.
In the future, prompt should be chosen by the user though.

PCData is now an object type that contains just player data.
For now it's things like your password, but in the future it will
have things like settings and configuration details.

SaveData is now the data structurre that represents a
fully saved character. It's essentially a mix of the PCData
fields (password) and a full representation of the saveable
details of the mob (name, location, race, class, etc).